requirements.txt
alembic==1.9.2
attrs==22.2.0
autopep8==2.0.1
click==8.1.3
exceptiongroup==1.1.0
flasgger==0.9.5
Flask==2.2.2
Flask-Migrate==4.0.2
Flask-SQLAlchemy==3.0.2
Flask-Testing==0.8.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
mistune==2.0.4
packaging==23.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyrsistent==0.19.3
pytest==7.2.1
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.46
tomli==2.0.1
Werkzeug==2.2.2
